move browse and image to subpackage of handler.
